Add support (with tests) for read-only web resource sets.
git-svn-id: https://svn.apache.org/repos/asf/tomcat/trunk@1558395 13f79535-47bb-0310-9956-ffa450edef68

+++ java/org/apache/catalina/webresources/FileResource.java
@@ -42,9 +42,10 @@ public class FileResource extends AbstractResource {
 
     private final File resource;
     private final String name;
+    private final boolean readOnly;
 
     public FileResource(WebResourceRoot root, String webAppPath,
-            File resource) {
+            File resource, boolean readOnly) {
         super(root,webAppPath);
         this.resource = resource;
 
@@ -64,6 +65,8 @@ public class FileResource extends AbstractResource {
             // Must be a file
             name = resource.getName();
         }
+
+        this.readOnly = readOnly;
     }
 
     @Override
@@ -93,6 +96,9 @@ public class FileResource extends AbstractResource {
 
     @Override
     public boolean delete() {
+        if (readOnly) {
+            return false;
+        }
         return resource.delete();
     }
